pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-portfolio-manager"
version = "0.1.0"
description = "Terminal-based stock portfolio analysis tool powered by Yahoo Finance and Claude AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["stock", "portfolio", "finance", "analysis", "cli", "terminal", "ai", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "yfinance>=0.2.40,<1.0",
    "anthropic>=0.40.0",
    "rich>=14.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
portfolio = "app.cli:main"
stock-agent = "app.cli:main"
portfolio-newsletter = "app.newsletter:main"

[project.urls]
Homepage = "https://github.com/mukul/terminal-portfolio-manager"
Repository = "https://github.com/mukul/terminal-portfolio-manager"
Issues = "https://github.com/mukul/terminal-portfolio-manager/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
"*" = ["*.env.example"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
